+++ lua/rustaceanvim/commands/expand_macro.lua
@@ -2,11 +2,6 @@ local ui = require('rustaceanvim.ui')
 
 local M = {}
 
----@return lsp_position_params
-local function get_params()
-  return vim.lsp.util.make_position_params()
-end
-
 ---@type integer | nil
 local latest_buf_id = nil
 
@@ -70,11 +65,15 @@ local function handler(_, result)
   ui.resize(true, '-25')
 end
 
-local rl = require('rustaceanvim.rust_analyzer')
-
 --- Sends the request to rust-analyzer to expand the macro under the cursor
 function M.expand_macro()
-  rl.buf_request(0, 'rust-analyzer/expandMacro', get_params(), handler)
+  local ra = require('rustaceanvim.rust_analyzer')
+  local clients = ra.get_active_rustaceanvim_clients(0)
+  if #clients == 0 then
+    return
+  end
+  local params = vim.lsp.util.make_position_params(0, clients[1].offset_encoding or 'utf-8')
+  ra.buf_request(0, 'rust-analyzer/expandMacro', params, handler)
 end
 
 return M.expand_macro