+++ lua/rustaceanvim/commands/hover_range.lua
@@ -42,18 +42,17 @@ local function get_visual_selected_range()
   return make_lsp_position(row1, math.min(col1, col2), row1, math.max(col1, col2))
 end
 
----@return lsp_range_params
-local function get_opts()
-  local params = vim.lsp.util.make_range_params()
+function M.hover_range()
+  local ra = require('rustaceanvim.rust_analyzer')
+  local clients = ra.get_active_rustaceanvim_clients(0)
+  if #clients == 0 then
+    return
+  end
+  local params = vim.lsp.util.make_range_params(0, clients[1].offset_encoding or 'utf-8')
+  ---@diagnostic disable-next-line: inject-field
   params.position = get_visual_selected_range()
   params.range = nil
-  return params
-end
-
-local rl = require('rustaceanvim.rust_analyzer')
-
-function M.hover_range()
-  rl.buf_request(0, 'textDocument/hover', get_opts())
+  ra.buf_request(0, 'textDocument/hover', params)
 end
 
 return M.hover_range