+++ lua/rustaceanvim/hover_actions.lua
@@ -3,10 +3,6 @@ local lsp_util = vim.lsp.util
 
 local M = {}
 
-local function get_params()
-  return lsp_util.make_position_params(0, nil)
-end
-
 ---@class rustaceanvim.hover_actions.State
 local _state = {
   ---@type integer
@@ -30,10 +26,8 @@ end
 -- run the command under the cursor, if the thing under the cursor is not the
 -- command then do nothing
 ---@param ctx table
-local function run_command(ctx)
-  local winnr = vim.api.nvim_get_current_win()
-  local line = vim.api.nvim_win_get_cursor(winnr)[1]
-
+---@param line integer
+local function run_command(ctx, line)
   if line > #_state.commands then
     return
   end
@@ -61,6 +55,7 @@ local function parse_commands()
   return prompt
 end
 
+---@param ctx lsp.HandlerContext
 function M.handler(_, result, ctx)
   if not (result and result.contents) then
     -- return { 'No information available' }
@@ -132,15 +127,24 @@ function M.handler(_, result, ctx)
 
   -- run the command under the cursor
   vim.keymap.set('n', '<CR>', function()
-    run_command(ctx)
+    local line = vim.api.nvim_win_get_cursor(winnr)[1]
+    run_command(ctx, line)
   end, { buffer = bufnr, noremap = true, silent = true })
+  vim.keymap.set('n', '<Plug>RustHoverAction', function()
+    local line = math.max(vim.v.count, 1)
+    run_command(ctx, line)
+  end, { buffer = vim.api.nvim_get_current_buf(), noremap = true, silent = true })
 end
-
-local rl = require('rustaceanvim.rust_analyzer')
 
 --- Sends the request to rust-analyzer to get hover actions and handle it
 function M.hover_actions()
-  rl.buf_request(0, 'textDocument/hover', get_params(), M.handler)
+  local ra = require('rustaceanvim.rust_analyzer')
+  local clients = ra.get_active_rustaceanvim_clients(0)
+  if #clients == 0 then
+    return
+  end
+  local params = lsp_util.make_position_params(0, clients[1].offset_encoding or 'utf-8')
+  ra.buf_request(0, 'textDocument/hover', params, M.handler)
 end
 
 return M