+++ lua/rustaceanvim/neotest/init.lua
@@ -273,6 +273,7 @@ function NeotestAdapter.build_spec(run_args)
     local dap = require('rustaceanvim.dap')
     overrides.sanitize_command_for_debugging(runnable.args.cargoArgs)
     local future = nio.control.future()
+    ---@diagnostic disable-next-line: invisible
     dap.start(runnable.args, false, function(strategy)
       future.set(strategy)
     end, function(err)
@@ -297,7 +298,6 @@ function NeotestAdapter.build_spec(run_args)
   local is_cargo_test = args[1] == 'test'
   local insert_pos = is_cargo_test and 2 or 3
   table.insert(args, insert_pos, '--no-fail-fast')
-  table.insert(args, insert_pos, '--color=never')
   if is_cargo_test then
     -- cargo test needs to pass --color=never to the test runner too
     table.insert(args, '--color=never')